feat(uffs-mft): offline $Bitmap decode + split metafile_decode module (P6)
- parse_bitmap: decode cluster-allocation stats (total/used/free clusters) from
  a captured $Bitmap payload by popcounting the allocation bitmap; folded into
  summarize / `metafile-info` for offline free-space validation (vs fsutil).
- Extract the pure offline decoders (BootGeometry/parse_boot, BitmapStats/
  parse_bitmap, summarize) into a new cross-platform `platform::metafile_decode`
  module — the $Bitmap addition pushed metafile.rs over the file-size limit, so
  this is the real fix (metafile.rs 826 → 661 LOC), no exception.

Cross-platform + unit-tested; verified end-to-end on macOS. host pedantic +
cargo-xwin Windows clippy clean.
